refactor(package): create a new package ui to store all user intefaces
