--- .env
@@ -1,3 +1,0 @@
-TELEGRAM_WEBHOOKS=https://integram.org/ckNa3mjYlqw
-GIT_REPO=flyve-mdm/flyve-mdm-android-agent
-GIT_BRANCH=feature/tag
