+++ app/src/main/AndroidManifest.xml
@@ -24,13 +24,13 @@
     <uses-permission android:name="android.permission.CAMERA" />
 
     <application
-        android:name=".MDMAgent"
+        android:name=".ui.MDMAgent"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".SplashActivity"
+            android:name=".ui.SplashActivity"
             android:configChanges="orientation|screenSize"
             android:launchMode="singleTask"
             android:screenOrientation="portrait"
@@ -42,31 +42,31 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:configChanges="orientation|screenSize">
         </activity>
         <activity
-            android:name=".EnrollmentActivity"
+            android:name=".ui.EnrollmentActivity"
             android:configChanges="orientation|screenSize">
         </activity>
         <activity
-            android:name=".EditUserActivity"
+            android:name=".ui.EditUserActivity"
             android:configChanges="orientation|screenSize">
         </activity>
         <activity
-            android:name=".PreviewSupervisorActivity"
+            android:name=".ui.PreviewSupervisorActivity"
             android:configChanges="orientation|screenSize">
         </activity>
         <activity
-            android:name=".DisclosureActivity"
+            android:name=".ui.DisclosureActivity"
             android:configChanges="orientation|screenSize">
         </activity>
         <activity
-            android:name=".PreviewUserActivity"
+            android:name=".ui.PreviewUserActivity"
             android:configChanges="orientation|screenSize">
         </activity>
         <activity
-            android:name=".StartEnrollmentActivity"
+            android:name=".ui.StartEnrollmentActivity"
             android:configChanges="orientation|screenSize"
             android:launchMode="singleTask"
             >
