+++ app/src/main/java/org/flyve/mdm/agent/ui/MDMAgent.java
@@ -25,7 +25,7 @@
  * ------------------------------------------------------------------------------
  */
 
-package org.flyve.mdm.agent;
+package org.flyve.mdm.agent.ui;
 
 
 import android.app.Application;
