--- ci/gplay.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "api-5742136459282318571-916118",
-  "private_key_id": "dde12ab6d2bdfdb2641aeea4cd903779841ec7ab",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCWK2bgJO+3CWQ4\nENjk4eoN/23QWtDQAJkzUGC1lU1+/UAdDE41XhqMJjoQPLEHiVrOsaidR8p3xycL\nvswnraeygXfjJOTTNEhbMtWN9f3sauMvvOm+0aPJpVE4bsmRPhWDM7pP1hID34g+\npHXwLZOR/MDKnvzWNWZbsiQNuZj5h4j4iaPPY+fGTH2UYLuN9d65NIbhRLKf+wcJ\nocZYGueqzZnwdAGI0X0o9OpjSY5CWm0NkAEdSpQce+CoOK4fB07tjtul6V7yZsMM\nG8yJt/oS6XfC+uOaH9/DqUzdTTwcwsAVUB6D1PD6e85wUs+vjgmHjvwsEsX6LVaN\nC/3KfNtdAgMBAAECggEAIy+wPzzoXj75GJn5+EdnDDiZhoSBAPvi6b8fqex96A93\nAldry91s4wHiAQP3hwiMGnpOQ+waaM5lO79NHs71yEHF0ZdHdqV9fsD+gPDuuoKh\nEQJ20ptUgV8XFN4ApOOjLsbkFOjX98hcqDDo4YoMVdyfX9q464g/n6SyBfVC/tSD\nM5pR0FTUOAj/5VoQNPO27xYbWnLVHz1Gf5Vww8RmM721GYYDzX9VEILQ9pHawwDw\nVWKDXKQ3aWi5YXCf0PKyqetHLtCMjVsZZaBmkNcBiXf8a1Nu3kpoKZO0q6QvA5+i\nJCkYUWEquKdjNl/F1hmMwzM8bxXaE3g98MK79BlWgQKBgQDL1UoOS6GEwzIB3DOX\nBbv49OtP16VoLkHvfxDRn8XSFckR894PAgD9vathRTCFGJttXQl4/Fz/tWu2+gnE\n4qHls0Pg2Wbwn7gPA+uxsuJNI+MNAuX+wWsmdhRkAsDJOjZGDIVs4KA06okTkbbN\nCpI2gHB/wR5y31CD2poTVvu3JQKBgQC8mi/01CLekC/Bss8XnKH+QoBjga8JDzDT\n17h+ZwZuVhHJjMV4kdTsQALkhK1zxsq/CLN23x4xBCsfkzWgDqz4MB/iKFZ/KYtQ\nmA/kq8wZnumC5sLE8VuX0/PDdHDuQO3fZcMVfSvdLIU8rIvvm7mPMpjbeHQBVAnx\n0mAZaowZ2QKBgGgEfDo3VkGN6uMQIDOddYNiWQLxpTM5B+TUDx0qT6UaAPXsYRF/\nkEnslJya1lvVZtu0crUO89Z6GnHRv1J95Oh3RthjmSiKGT9wXMLd6qY6/g8Rv5T1\n11xVPaqbzQ3javbsRNp+9sKKPVGDNqoEDwFJIh/PyV4Nn+TplV7xMJsRAoGBAIfo\nLAqnG4F6MY/JREZljq5kSZRrDTDgC3V8i+w6XX4A88xjbqFVpTOhZPQ7T6qQY+WJ\nZZmRanYz5MdaE1iFestWEoOsQbhcvZGUIk/MJMWJqAe2NkApFLGyAISmUgK+Myu6\nbvv7OTKmNFTmmehONUclVNVCVYXng0sJKEroZpDxAoGAKO0EhrHoPrKck0s5tBD6\ng1T/p44BEjdnKI/XpoQKX/xiWGbP5XIuz9TA22xs1kdAMR50AJl0JXkJg9YR/ROv\n9oNDDEeZH6XTLNPlOYCEQlGte13E/MP09bxxWJUXlpex0MquY9gOD1twuQQxdEsl\nvTSi+nRECyfnViT143IYYCk=\n-----END PRIVATE KEY-----\n",
-  "client_email": "service-account@api-5742136459282318571-916118.iam.gserviceaccount.com",
-  "client_id": "117597446770034169545",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://accounts.google.com/o/oauth2/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/service-account%40api-5742136459282318571-916118.iam.gserviceaccount.com"
-}
